+++ tests/cmake/AddTestMaps.cmake
@@ -199,6 +199,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1617_crash_disabled_chest"
   "bugs/1619_camera_obstacles"
   "bugs/1628_hero_set_direction_crash"
+  "bugs/1635_timer_set_remaining_time"
 )
 
 # Test that should succeed when there are non-fatal errors