+++ changelog.md
@@ -51,6 +51,7 @@ The multiplayer update!
 * Add a method `game:get_values()` (#1552).
 * Add a method `movement:get_type()` (#840).
 * Allow to disable the spin attack (#771).
+* Add methods `timer:get/set_delay()`.
 
 ### Solarus launcher GUI changes